+++ pandas/tests/frame/test_analytics.py
@@ -891,7 +891,7 @@ class TestDataFrameAnalytics(TestData):
         mixed['F'] = Timestamp('20130101')
 
         # results in an object array
-        from pandas.core.tools.timedeltas import (
+        from pandas._libs.tslibs.timedeltas import (
             _coerce_scalar_to_timedelta_type as _coerce)
 
         result = mixed.min()
