+++ pandas/tests/scalar/test_timedelta.py
@@ -6,10 +6,13 @@ from datetime import timedelta
 
 import pandas as pd
 import pandas.util.testing as tm
-from pandas.core.tools.timedeltas import _coerce_scalar_to_timedelta_type as ct
-from pandas import (Timedelta, TimedeltaIndex, timedelta_range, Series,
-                    to_timedelta, compat)
+
+from pandas import TimedeltaIndex, timedelta_range, Series, compat
 from pandas._libs.tslib import iNaT, NaT
+from pandas._libs.tslibs.timedeltas import (
+    Timedelta,
+    to_timedelta,
+    _coerce_scalar_to_timedelta_type as ct)
 
 
 class TestTimedeltaArithmetic(object):
